synthetic-agent/requirements.txt
fastapi
uvicorn
requests
openai
pydantic
sqlalchemy
pytest
psycopg2-binary
python-dotenv
